+++ jest.config.cjs
@@ -9,6 +9,7 @@ module.exports = {
   preset: 'ts-jest',
   testEnvironment: 'node',
   moduleDirectories: ['<rootDir>', 'node_modules'],
-  testRegex: '.*.(spec|test).ts$',
+  testRegex: '.*\.(spec|test)\.ts$',
   collectCoverageFrom: ['src/**/*.ts'],
+  passWithNoTests: true,
 };