+++ src/utils/path.ts
@@ -1,6 +1,5 @@
 import path from 'node:path';
 
 export const ROOT_PATH = process.cwd();
-export const PACKAGE_JSON_PATH = path.join(ROOT_PATH, 'package.json');
 
 export const rootPath = (...paths: string[]) => path.join(ROOT_PATH, ...paths);